transfer errorDetails to defined in constants.ts
